requirements.txt
flask==2.0.1
pymongo==4.1.1
python-dotenv==0.19.2
requests==2.27.1